Set lane and decoration prefabs to use layers
Assigned the grassLane, trackLane and roadLane the Ground layer. Assigned the waterLane and the LilypadLane the Water Layer. Removed the raycast and instead implemented the box cast for better collision detection, changed the boundary wall positons to 6 and -6. updated the camera follow script to include camera shake on death and movement, fixed the camera pitch and yaw, adjusted the camera position, added a particle system to to the player for a death effect, added coin and score text fonts, added player prefab.